+++ src/main/java/fr/eni/encheres/dal/RetraitDAOImpl.java
@@ -3,8 +3,9 @@ package fr.eni.encheres.dal;
 import fr.eni.encheres.bo.Retrait;
 
 public class RetraitDAOImpl implements RetraitDAO {
+
     @Override
-    public Retrait read(int idArticle) {
+    public Retrait read(long idArticle) {
         return null;
     }
